backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
click==8.1.7
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pluggy==1.4.0
psycopg2==2.9.9
pyarrow==15.0.0
pyclean==2.7.6
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0.post1
